add Sort component
